+++ App/plugins/Shape/Plane/include/RayTracer/Plane.hpp
@@ -20,9 +20,9 @@ namespace RayTracer {
 
             [[nodiscard]] std::string getPluginName() const override { return PLANE_SHAPE; };
 
-            void setRadius(const int16_t& radius) override { (void) radius; throw RunTimeException("Plane shape does not have a radius");};
+            void setRadius(const double& radius) override { (void) radius; throw RunTimeException("Plane shape does not have a radius");};
 
-            [[nodiscard]] int16_t getRadius() const override { throw RunTimeException("Plane shape does not have a radius");};
+            [[nodiscard]] double getRadius() const override { throw RunTimeException("Plane shape does not have a radius");};
             [[nodiscard]] Vector& getRotation() override { throw RunTimeException("Plane shape does not have a rotation");};
 
     }; // class Plane
